Modify Deploy logic for Prover Tests
