+++ contracts/Prover.sol
@@ -225,9 +225,13 @@ contract Prover is SimpleProver, AbstractProver {
      * state.
      */
     function proveSettlementLayerState(bytes calldata rlpEncodedBlockData) public {
-        console.log("Hello I'm in proveSettlementLayerState");
+        console.log("Hi I'm in proveSettlementLayerState");
+        console.log("block.chainid", block.chainid);
         uint256 settlementChainId = chainConfigurations[block.chainid][ProvingMechanism.Cannon].settlementChainId;
+        console.log("settlementChainId", settlementChainId);
+        console.log("First test");
         if (!chainConfigurations[settlementChainId][ProvingMechanism.Settlement].exists) {
+            console.log("About to revert");
             revert InvalidDestinationProvingMechanism(block.chainid, ProvingMechanism.Settlement);
         }
         require(keccak256(rlpEncodedBlockData) == l1BlockhashOracle.hash(), "hash does not match block data");
@@ -247,10 +251,13 @@ contract Prover is SimpleProver, AbstractProver {
             stateRoot: bytes32(RLPReader.readBytes(RLPReader.readList(rlpEncodedBlockData)[3]))
         });
         BlockProof memory existingBlockProof = provenStates[settlementChainId][SettlementType.Confirmed];
+        console.log("Second test");
         if (existingBlockProof.blockNumber < blockProof.blockNumber) {
+            console.log("Second test ok");
             provenStates[settlementChainId][SettlementType.Confirmed] = blockProof;
             emit L1WorldStateProven(blockProof.blockNumber, blockProof.stateRoot);
         } else {
+            console.log("Have Outdated Block");
             revert OutdatedBlock(blockProof.blockNumber, existingBlockProof.blockNumber);
         }
     }
@@ -280,6 +287,9 @@ contract Prover is SimpleProver, AbstractProver {
         uint256 l2settlementChainId = chainConfigurations[block.chainid][ProvingMechanism.Bedrock].settlementChainId;
         uint256 settlementChainId =
             chainConfigurations[l2settlementChainId][ProvingMechanism.Settlement].settlementChainId;
+        console.log("l2settlementChainId", l2settlementChainId);
+        console.log("settlementChainId", settlementChainId);
+        console.log("chainConfigurations", chainConfigurations[settlementChainId][ProvingMechanism.SettlementL3].exists);
         if (!chainConfigurations[settlementChainId][ProvingMechanism.SettlementL3].exists) {
             revert InvalidDestinationProvingMechanism(settlementChainId, ProvingMechanism.SettlementL3);
         }
@@ -491,6 +501,8 @@ contract Prover is SimpleProver, AbstractProver {
         FaultDisputeGameProofData memory faultDisputeGameProofData,
         bytes32 l1WorldStateRoot
     ) public {
+        console.log("Hi I'm in proveWorldStateCannon");
+        console.log("chainId", chainId);
         if (!chainConfigurations[chainId][ProvingMechanism.Cannon].exists) {
             revert InvalidDestinationProvingMechanism(chainId, ProvingMechanism.Cannon);
         }
