+++ package.json
@@ -147,6 +147,8 @@
     "@hyperlane-xyz/sdk": "^5.1.0",
     "@openzeppelin/contracts": "^5.0.2",
     "@openzeppelin/contracts-upgradeable": "^5.0.2",
-    "loadash": "^1.0.0"
+    "@types/lodash": "^4.17.12",
+    "loadash": "^1.0.0",
+    "lodash": "^4.17.21"
   }
 }
